refacto(highlighting): refactoring of the highlighting functions
This commit improves the code quality of the highlighting functions.
The refactoring was to divise the previous huge highlighting function
into small functions that will implement the logic to highlight
numbers, comments, strings and chars.

Important note: There are at least two small issues with the code that
will need to be fixed in the future.

The loop in highligh implicitly relies on the highlight function to
advance index. If any of these functions returns true, but does not
modify index, we will run into an inifinite loop.

The second issue is that we are pushing things into the highlighting
array without checking if it is safe, and we are advancing index in
many cases without any kind of check. One small fix for this issue
was to implement the  method to increase the array
capacity at least one more before pushing to it.